+++ seanime-desktop/src-tauri/src/tray.rs
@@ -1,3 +1,4 @@
+use crate::constants::MAIN_WINDOW_LABEL;
 use tauri::{
     menu::{Menu, MenuItem},
     tray::{MouseButton, MouseButtonState, TrayIconBuilder, TrayIconEvent},
@@ -6,7 +7,10 @@ use tauri::{
 
 pub fn create_tray<R: Runtime>(app: &tauri::AppHandle<R>) -> tauri::Result<()> {
     let quit_i = MenuItem::with_id(app, "quit", "Quit Seanime", true, None::<&str>)?;
-    let menu = Menu::with_items(app, &[&quit_i])?;
+    let toggle_visibility_i = MenuItem::with_id(app, "toggle_visibility", "Toggle visibility", true, None::<&str>)?;
+
+
+    let menu = Menu::with_items(app, &[&toggle_visibility_i, &quit_i])?;
 
     let _ = TrayIconBuilder::with_id("tray")
         .icon(app.default_window_icon().unwrap().clone())
@@ -15,6 +19,16 @@ pub fn create_tray<R: Runtime>(app: &tauri::AppHandle<R>) -> tauri::Result<()> {
         .on_menu_event(move |app, event| match event.id.as_ref() {
             "quit" => {
                 app.exit(0);
+            }
+            "toggle_visibility" => {
+                if let Some(window) = app.get_webview_window(MAIN_WINDOW_LABEL) {
+                    if window.is_minimized().unwrap() {
+                        let _ = window.show();
+                        let _ = window.set_focus();
+                    } else {
+                        let _ = window.minimize();
+                    }
+                }
             }
             // Add more events here
             _ => {}
@@ -27,7 +41,7 @@ pub fn create_tray<R: Runtime>(app: &tauri::AppHandle<R>) -> tauri::Result<()> {
             } = event
             {
                 let app = tray.app_handle();
-                if let Some(window) = app.get_webview_window("main") {
+                if let Some(window) = app.get_webview_window(MAIN_WINDOW_LABEL) {
                     let _ = window.show();
                     let _ = window.set_focus();
                 }